requirements.txt
flask==2.3.3
python-dotenv==1.0.0
google-generativeai==0.3.2
whisper==1.1.10
ffmpeg-python==0.2.0
openai-whisper==20231117
pymupdf==1.23.8
python-pptx==0.6.23
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0 